refactor: make protocol a single version
